renamed the configurations file and namespace to 'dynamic_parameters' to be more explicit
